azure-document-intelligence/requirements.txt
azure-ai-documentintelligence
azure-core
python-dotenv
